pyproject.toml
[project]
name = "bash-skills-agent"
version = "0.1.0"
description = "Google ADK-based agent application"
requires-python = ">=3.12"
readme = "README.md"
authors = [
    {name = "Yunju Choi", email = "hnclovera@gmail.com"}
]
dependencies = [
    "google-adk>=1.22.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=7.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.15.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bash_skills_agent"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
